Add set std fn to directly set a value to memory
It should be a dereference and assignment (*ptr = c), but we want a high level
language that can deal with strings as first class objects.
Set is a plumbing function that should not be used directly by the user.
